article-website/requirements.txt
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.0
Flask-Login==0.5.0
markdown==3.3.7
python-dotenv==0.19.2
Werkzeug==2.2.2